fix overflow

+++ OpenNos.DAL.EF.MySQL/DB/OpenNos.edmx.sql
@@ -44,9 +44,9 @@
 -- -----------------------------------------------------------
 -- Entity Designer DDL Script for MySQL Server 4.1 and higher
 -- -----------------------------------------------------------
--- Date Created: 04/04/2016 23:29:09
+-- Date Created: 04/05/2016 16:17:42
 
--- Generated from EDMX file: C:\Users\ERWAN\Desktop\OpenNos Git\OpenNos.DAL.EF.MySQL\DB\OpenNos.edmx
+-- Generated from EDMX file: C:\Users\Dominik\Source\Repos\OpenNos\OpenNos.DAL.EF.MySQL\DB\OpenNos.edmx
 -- Target version: 3.0.0.0
 
 -- --------------------------------------------------
@@ -411,7 +411,7 @@ ALTER TABLE `npcmonster` ADD PRIMARY KEY (NpcMonsterVNum);
 CREATE TABLE `inventoryitem`(
 	`InventoryItemId` bigint NOT NULL AUTO_INCREMENT UNIQUE, 
 	`ItemVNum` smallint NOT NULL, 
-	`Amount` smallint NOT NULL, 
+	`Amount` int NOT NULL, 
 	`Rare` TINYINT UNSIGNED NOT NULL, 
 	`Upgrade` TINYINT UNSIGNED NOT NULL, 
 	`Design` smallint NOT NULL,
